Added NomialData class
