+++ gpkit/geometric_program.py
@@ -2,13 +2,10 @@
 import numpy as np
 
 from time import time
-from functools import reduce as functools_reduce
-from operator import add
 
 from .variables import Variable, VectorVariable
 from .small_classes import CootMatrix
-from .small_scripts import locate_vars
-from .small_scripts import mag
+from .nomial_data import NomialData
 
 
 class GeometricProgram(object):
@@ -28,8 +25,7 @@ class GeometricProgram(object):
 
     Attributes with side effects
     ----------------------------
-    `solver_out` is set during a solve
-    `solver_log` is set during a solve
+    `solver_out` and `solver_log` are set during a solve
     `result` is set at the end of a solve
 
     Examples
@@ -47,11 +43,11 @@ class GeometricProgram(object):
         self.cost = cost
         self.constraints = constraints
         self.posynomials = [cost] + list(constraints)
-        self.cs = np.hstack((mag(p.cs) for p in self.posynomials))
-        if not all(self.cs > 0):
+        nd = NomialData(signomials=self.posynomials, simplify=False)
+        if not all(nd.cs > 0):
             raise ValueError("GeometricPrograms cannot contain Signomials.")
-        self.exps = functools_reduce(add, (p.exps for p in self.posynomials))
-        self.varlocs, self.varkeys = locate_vars(self.exps)
+        self.exps, self.cs = nd.exps, nd.cs
+        self.varlocs, self.varstrs = nd.varlocs, nd.varstrs
         # k [j]: number of monomials (columns of F) present in each constraint
         self.k = [len(p.cs) for p in self.posynomials]
         # p_idxs [i]: posynomial index of each monomial
@@ -144,7 +140,7 @@ class GeometricProgram(object):
         if "objective" in solver_out:
             result["cost"] = float(solver_out["objective"])
         else:
-            result["cost"] = self.cost.subcmag(result["variables"])
+            result["cost"] = self.cost.subsummag(result["variables"])
 
         result["sensitivities"] = {}
         if "nu" in solver_out:
