+++ gpkit/substitution.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 
-from collections import defaultdict
+from collections import defaultdict, Iterable
 
 from .small_classes import Numbers, Strings, Quantity
 from .small_classes import HashVector
@@ -17,9 +17,9 @@ from .small_scripts import mag
 from . import DimensionalityError
 
 
-def getsubs(varkeys, varlocs, substitutions):
+def getconstants(nomial, substitutions):
     subs = {}
-    varset = frozenset(varlocs)
+    varset = frozenset(nomial.varlocs)
     for var, sub in substitutions.items():
         if not is_sweepvar(sub):
             if isinstance(var, Monomial):
@@ -27,8 +27,8 @@ def getsubs(varkeys, varlocs, substitutions):
                 if var_ in varset:
                     subs[var_] = sub
             elif isinstance(var, Strings):
-                if var in varkeys:
-                    var_ = varkeys[var]
+                if var in nomial.varstrs:
+                    var_ = nomial.varstrs[var]
                     vectorsub(subs, var_, sub, varset)
             else:
                 vectorsub(subs, var, sub, varset)
@@ -70,7 +70,40 @@ def vectorsub(subs, var, sub, varset):
         subs[var] = sub
 
 
-def substitution(varlocs, varkeys, exps, cs, substitutions, val=None):
+def separatesubs(nomial, substitutions):
+    "Separate substitutions by type."
+    sweep, linkedsweep = {}, {}
+    constants = dict(substitutions)
+    for var, sub in substitutions.items():
+        if is_sweepvar(sub):
+            del constants[var]
+            if isinstance(var, Strings):
+                var = nomial.varstrs[var]
+            elif isinstance(var, Monomial):
+                var = VarKey(var)
+            if isinstance(var, Iterable):
+                suba = np.array(sub[1])
+                if len(var) == suba.shape[0]:
+                    for i, v in enumerate(var):
+                        if hasattr(suba[i], "__call__"):
+                            linkedsweep.update({VarKey(v): suba[i]})
+                        else:
+                            sweep.update({VarKey(v): suba[i]})
+                elif len(var) == suba.shape[1]:
+                    raise ValueError("whole-vector substitution"
+                                     " is not yet supported")
+                else:
+                    raise ValueError("vector substitutions must share a"
+                                     "dimension with the variable vector")
+            elif hasattr(sub[1], "__call__"):
+                linkedsweep.update({var: sub[1]})
+            else:
+                sweep.update({var: sub[1]})
+    constants = getconstants(nomial, constants)
+    return sweep, linkedsweep, constants
+
+
+def substitution(nomial, substitutions, val=None):
     """Efficient substituton into a list of monomials.
 
         Arguments
@@ -101,19 +134,20 @@ def substitution(varlocs, varkeys, exps, cs, substitutions, val=None):
     if val is not None:
         substitutions = {substitutions: val}
 
-    subs = getsubs(varkeys, varlocs, substitutions)
+    subs = getconstants(nomial, substitutions)
 
     if not subs:
-        return varlocs, exps, cs, subs
+        return nomial.varlocs, nomial.exps, nomial.cs, subs
         # raise KeyError("could not find anything to substitute"
         #                "in %s" % substitutions)
 
-    exps_ = [HashVector(exp) for exp in exps]
-    cs_ = np.array(cs)
+    exps_ = [HashVector(exp) for exp in nomial.exps]
+    cs_ = np.array(nomial.cs)
     varlocs_ = defaultdict(list)
-    varlocs_.update({var: list(idxs) for (var, idxs) in varlocs.items()})
+    varlocs_.update({var: list(idxs)
+                     for (var, idxs) in nomial.varlocs.items()})
     for var, sub in subs.items():
-        for i in varlocs[var]:
+        for i in nomial.varlocs[var]:
             x = exps_[i].pop(var)
             varlocs_[var].remove(i)
             if len(varlocs_[var]) == 0:
